(tests) Tag Internal tests that require official
Some of the new tab completions utilize choco.exe to determine what is
available for completion. These tests can only work in a scenario where
Chocolatey CLI is built using the official keys. As such, the tests for
this completion have been updated to be within a Context block that is
tagged `Internal` so that they are excluded when run locally.