(#3672) CREDITS 2.0.0
This contains updates to the CREDITS files with the
new global format. The CREDIT.md file has been
updated and all of the other info in it has been
moved to the THANKS.md file. There is also a new
CREDITS.json file that contains information about
all the dependencies, as well as a PDF version for
easy reading.