(maint) Update GitReleaseManager template
As part of the CLI 2.5.0 release, we made use of the GitReleaseManager
template that includes information about contributors to the release.
You can see the output here:

https://github.com/chocolatey/choco/releases/tag/2.5.0

We as a team are very happy with this output, however, the output that
you see there is not what was generated by GitReleaseManager, it needed
a couple of tweaks.

This commit makes changes to the contributor template so that the
manual tweaks will not be required going forward.

These tweaks were:

- Move the information about the issue number and who solved the
problem in a PR, to it's own line.  The single line approach was
getting very "busy" and it was decided that a new line made a lot of
sense
- The title for "Contributors" was updated to use `###` rather than
surrounding with '__'.  This was just so that things were consistent in
the generated file.
- The if statement for including the milestone was corrected to only do
the inclusion is the milestone description actually has some content.
This change prevented the large gap that was left at the start of the
file
- General spacing around each issue section was updated to remove empty
lines
- After breaking the content onto two lines, "see" was changed to "See"
since it was a start of a new line.  Also a "." was added to the end of
the first line
- Remove usage of the __ on the issues description. This doesn't really
have much impact on the output, and doesn't add anything, so let's
simply not use it.
- Finally, an extra "," was introduced to space out the different
sections of the section line, making it easier to read.