(#3672) Add "info" to default license links
For some dependencies, they do not contain a tag or a version in the
repository. Instead of linking to the license of master or main, we now
link to the permalink to better track history over time. We do not want
this going to master or main incase a license type changes in the future
and the dependency becomes inaccurate. An "info" entry has been added
as well, to explain this change further.