(tests) Add Pester tests to cover searching CCR
There are a number of additional Chocolatey CLI command line options
that are only intended to work when Chocolatey Community Repository is
being used.  Previously, these options were not tested, as the
infrastructure to exercise these didn't exist.  That is no longer the
case, and it is now possible to test the following command line options:

- --download-cache
- --not-broken
- --approved-only
- --id-starts-with
- --by-id-only
- --by-tags-only
- --order-by-popularity
- --order=""

and to make assertions on the results, based on a known collection of
packages existing in the repository.

The Tag that is applied to each of these tests to allow exectuion of
these tests only at certain times, and not as part of the default
execution of these tests.