(tests) Improve test reliability
This commit takes some steps to improve overall test reliability.
Mainly:

- Do not copy the entire `lib` directory when creating a new
ChocolateyTestInstall. This allows us to then bring over only the
Chocolatey packages that we want for use in tests. It was discovered
that in some scenarios the packages were many tens of thousands of
files (thanks node_modules), and while it didn't produce errors, when
test failures happened there was a lot of output for things unrelated
to the tests.
- Clear the Chocolatey HTTP cache when a new snapshot is generated so
the new set of tests don't operate on old HTTP data. Again, this didn't
produce errors, but when tests failed they sometimes would mention
using the cache making troubleshooting more difficult as you try to
identify if it's the cache or the test.
- Remove times where we were setting $Output.Lines to $Output.Lines.
- Add an AfterAll to Remove the Chocolatey Install Snapshot where the
configuration is adjusted. This didn't result in test failures, but the
output would mention these changes that were unrelated to the test in
question and could lead to unnecessary investigations.
- Remove the `cory` tag from `choco-removed.Tests.ps1`... This tag was
used for testing, and was unintentionally left in.
-