(tests) Update the push test to pass CCR
The tests added for CCR, inadvertently used the `/api/v2/` endpoint for
pushing, this is incorrect, and should have just used the `/` endpoint.
Further, the output is different depending on if you're pushing to the
local CCR source, or if you're pushing to CCR itself. The validation of
that push has been updated to match as much of the line as will be
available.