(tests) Move the push test's source declaration
When adding #3774, we added a clearing of the HTTP cache when new test
snapshots are created. This has resulted in the push tests failing as
the cache no longer exists, and so it was prompting for credentials in
only one of the scenarios. This moves it so that the source is
configured for both of the scenarios.[