(build) Pin to macos-13 build agent
For now, let's pin to an earlier version of Mac OS.  There are LOTs of
changes on the Mac eco-system, which Chocolatey has not yet been
tested with, and with the newer versions of the Mac OS build agents
removing the installation of mono, it makes sense to pin to this version.  Since we are doing this, the manual installation of Mono is
no longer required.