(build) Pin to macos-14
Looking back through the build logs, it would appear that the build
was working fine on Mac OS 14.7.6:

https://github.com/chocolatey/choco/actions/runs/17250402137/job/48951031483

So going to attempt to run the build on there, since it is currently
failing on macos-13.  Rather than dig into what is going on with
the combination of git, and gitversion, let's try to run it where we
know it _was_ working.