(build) Remove usage of backports repository
This repository no longer exists, and is causing a build failure when
trying to build the docker image. The hope with this commit is that
the installed version of git on the base image is now high enough
that it is no longer causing the problems that it was before, and as a
result, the installation of a newer version of git is no longer required.