(#3756) Update dockerfile.linux to new base image
This commit is aiming to address a number of problems that we are
seeing with the current Docker build.

The first is that in order for the build to succeed, `git` needs to be available.  This was achieved previously by using the backports
repository and going `apt-get intstall git`, however, with the 
backports repository going away:

https://backports.debian.org/news/Removal_of_buster-backports_from_the_debian_archive/

this approach is no longer possible.

In addition the mono base image that we were using is also 
deprecated, and no longer being updated:

https://github.com/mono/docker

So we can't just update to a newer version.

Instead, we need to go the other way, namely, update the base
image being used directly.  Currently bookworm is LTS until 2028
so let's start there.  With that in place, we can install mono
directly, rather than already having it installed.  The steps for
doing this were borrowed from here:

https://cloudspinx.com/installing-mono-microsofts-net-framework-on-debian/

Finally, all original references to mono were removed from the
file, including the mention of it as the base name.

The installtion of curl is not directly required from a mono point
of view, for one of the first things that the build.sh does is to
fetch things using curl, so it does need to present.