(build) Remove setting of environment variable
The execution of docker build is currently failing when running in
GitHub Actions.  The same docker build works fine when running locally.
We suspect that GitVersion, which is the line that is failing, is
trying to do _something_ when the GITHUB_ACTIONS environment variable
is in play.  Since we don't actually care about GitVersion asserting
anything in addition to the actual build number at this point in the
build, let's simply remove the setting of this environment variable.