(build) Pin to specific Mac OS for tests build
This is similar to what has been done on the main build.  With GitHub
Actions switching to macos-15, there are known complications with
running the build there, as mono isn't currently installed on that.
For now, let's pin to mac0s-14, where we know the build is running, and
come back to get things working on macos-15 later.

Also, took the opportunity to update the docs for the docker build,
since the input argument for the Dockerfile no longer exists, since we
are no longer using the base Mono image.