(build) Consolidate changes to build on develop
As part of the preparation work for the next 1.x and 2.x releases, a
number of changes were made to the build infrastructure for GitHub
Actions, in order to get the Docker Build working, as well as fixing
the integration tests, and Mac builds.

This commit is a consolidation of all of those changes, including:

- Install Mono for latest ubuntu-latest image
- Removal of commented out sections
- Pin to macos-14, as macos-15 no longer has mono and the initial
attempt to simply install it, caused a build failure, so just pinned to
what works for now
- Removal of unnecessary build arguments being passed through to Docker
build
- Removal of caching of files for Docker build. This was found to be
causing the main problem with getting a working build, and it was
decided that caching for this build is not a requirement
- Switch to a new base image for Docker build (which required
installing some missing components), as well as tidying up the linting
warnings on the Dockerfile.