(build) Attempt to use reference assemblies
After bumping to windows-latest, the GitHub Actions build started
failing with this error:

The reference assemblies for .NETFramework,Version=v4.0 were not found.
To resolve this, install the Developer Pack (SDK/Targeting Pack) for
this framework version or retarget your application.

As a first pass, let's use the Reference Assemblies NuGet package.  If
this doesn't work, then we will need to look to installing the full
Developer Pack.