(maint) Chocolatey.PowerShell debugging directions
In order to debug the Chocolatey.PowerShell module in PowerShell, you
need to launch powershell.exe from Visual Studio, and then import the
module into that session. Visual Studio stores the project startup
information in the .csproj.user file which we explicitly ignore as
the user files are for the user's personal settings. In lieu of
committing a user specific file, this commit instead adds a ReadMe to
the Chocolatey.PowerShell project and provides instructions for setting
up for debugging.