(#3228) Throw earlier if .Net 4.8 is not present
There are instances where .Net is not installed in a way that allows
Chocolatey CLI to function. In these scenarios, the version in the
registry seems to be very old. This commit accounts for this by wrapping
the container registration in a try cath, and displaying a descriptive
error in that case.

Additionally, this changes the check to use Registry64 instead of
Registry32, as on 64 bit systems the 64 bit registry is the one that
determines the .Net version. And the lookup on 32 bit systems gracefully
falls back to the 32 bit registry.