(#3228) Update Pester Tests to validate .Net
This updates the .Net version Pester test to use the 64 bit registry as
that is the one that Chocolatey CLI is using. Further, it adds a test
for the .Net version to be a low integer instead of deleting it. It's
been found in some scenarios where the version is there but low .Net
misbehaves. It also adds a retry onto taking ownership of the registry
key as we've seen instances where it fails to take the ownership and so
the test fails. This retry should hopefully resolve these sporadic
failures.