(#3802) Add tests for licensed source
Ensure that when a license is installed, CLI adds the
chocolatey.licensed source correctly so that users can install the
licensed extension.

Additionally, when the license is removed, the feed should be removed,
and if the license is swapped out, CLI should remember the disabled
state of the licensed feed.

Noticed and removed a superfluous `return` statement in one of the
helper functions.