(#1901) Re-add display of deployment location
This commit reverts c74a19cdcf20942d3fda5383d821a78c1d6c6a4a where we
had reverted the initial work to display the deployment location. As we
are releasing a new version of CLE that can take this newer version as a
breaking change, we can add this functionality in there too.