(#3802) Avoid unnecessary writes to the config
If we don't avoid making changes unnecessarily, we can run into
situations where the config file is saved again every time it's read.

This can lead to odd behaviour if anything is watching the file to
monitor for changes, and also uses chocolatey.lib to read the config
file; this would cause a read->write->repeat loop.